Fix build and test on non-linux environments
